switch unsigned long to uint32_t

+++ common/net/utils_sockets.c
@@ -54,48 +54,48 @@ int get_udp_socket(const struct sockaddr *addr){
 }
 
 int recvSeqPacket(struct SeqSocket* ss, int packetLen) {
-   unsigned long  seq;
+   uint32_t	seq;
 
    if (packetLen < 0)
       return packetLen;
 
-   if (packetLen < (int)sizeof(unsigned long)) {
+   if (packetLen < (int)sizeof(uint32_t)) {
       if (seqErrorLogger)
-         seqErrorLogger(SEQ_noseq, ss->seqRecv, 0, ss->buffer - sizeof(unsigned long), (size_t)packetLen);
+         seqErrorLogger(SEQ_noseq, ss->seqRecv, 0, ss->buffer - sizeof(uint32_t), (size_t)packetLen);
 
       errno = EIO;
       return -1;
    }
 
-   seq = ntohl(((unsigned long*)ss->buffer)[-1]);
+   seq = ntohl(((uint32_t*)ss->buffer)[-1]);
 
    if (seq < ss->seqRecv) {
       if (seqErrorLogger)
-         seqErrorLogger(SEQ_backward, ss->seqRecv, seq, ss->buffer, (size_t)packetLen - sizeof(unsigned long));
+         seqErrorLogger(SEQ_backward, ss->seqRecv, seq, ss->buffer, (size_t)packetLen - sizeof(uint32_t));
 
       ss->seqRecv = seq + 1;
       errno = EIO;
       return -1;
    } else if (seq > ss->seqRecv) {
       if (seqErrorLogger)
-         seqErrorLogger(SEQ_skip, ss->seqRecv, seq, ss->buffer, (size_t)packetLen - sizeof(unsigned long));
+         seqErrorLogger(SEQ_skip, ss->seqRecv, seq, ss->buffer, (size_t)packetLen - sizeof(uint32_t));
    }
 
    ss->seqRecv = seq + 1;
 
-   return packetLen - sizeof(unsigned long);
+   return packetLen - sizeof(uint32_t);
 }
 
 int seq_read(struct SeqSocket* ss) {
-   return recvSeqPacket(ss, read(ss->socket, ss->buffer - sizeof(unsigned long), ss->maxSize + sizeof(unsigned long)));
+   return recvSeqPacket(ss, read(ss->socket, ss->buffer - sizeof(uint32_t), ss->maxSize + sizeof(uint32_t)));
 }
 
 int seq_recv(struct SeqSocket* ss, int flags) {
-   return recvSeqPacket(ss, recv(ss->socket, ss->buffer - sizeof(unsigned long), ss->maxSize + sizeof(unsigned long), flags));
+   return recvSeqPacket(ss, recv(ss->socket, ss->buffer - sizeof(uint32_t), ss->maxSize + sizeof(uint32_t), flags));
 }
 
 int seq_recvfrom(struct SeqSocket* ss, int flags, struct sockaddr* from, socklen_t* fromlen) {
-   return recvSeqPacket(ss, recvfrom(ss->socket, ss->buffer - sizeof(unsigned long), ss->maxSize + sizeof(unsigned long), flags, from, fromlen));
+   return recvSeqPacket(ss, recvfrom(ss->socket, ss->buffer - sizeof(uint32_t), ss->maxSize + sizeof(uint32_t), flags, from, fromlen));
 }
 
 int seq_send(struct SeqSocket* ss, size_t size, int flags) {
@@ -106,9 +106,9 @@ int seq_send(struct SeqSocket* ss, size_t size, int flags) {
       return -1;
    }
 
-   ((unsigned long*)ss->buffer)[-1] = htonl(ss->seqSend);
+   ((uint32_t*)ss->buffer)[-1] = htonl(ss->seqSend);
 
-   s = send(ss->socket, ss->buffer - sizeof(unsigned long), size + sizeof(unsigned long), flags);
+   s = send(ss->socket, ss->buffer - sizeof(uint32_t), size + sizeof(uint32_t), flags);
    if (s < 0)
       return s;
 
@@ -125,9 +125,9 @@ int seq_sendto(struct SeqSocket* ss, size_t size, int flags, const struct sockad
       return -1;
    }
 
-   ((unsigned long*)ss->buffer)[-1] = htonl(ss->seqSend);
+   ((uint32_t*)ss->buffer)[-1] = htonl(ss->seqSend);
 
-   s = sendto(ss->socket, ss->buffer - sizeof(unsigned long), size + sizeof(unsigned long), flags, to, tolen);
+   s = sendto(ss->socket, ss->buffer - sizeof(uint32_t), size + sizeof(uint32_t), flags, to, tolen);
    if (s < 0)
       return s;
 
@@ -154,9 +154,9 @@ int seq_write(struct SeqSocket* ss, size_t size) {
       return -1;
    }
 
-   ((unsigned long*)ss->buffer)[-1] = htonl(ss->seqSend);
+   ((uint32_t*)ss->buffer)[-1] = htonl(ss->seqSend);
 
-   s = write(ss->socket, ss->buffer - sizeof(unsigned long), size + sizeof(unsigned long));
+   s = write(ss->socket, ss->buffer - sizeof(uint32_t), size + sizeof(uint32_t));
    if (s < 0)
       return s;
